refactor(useReceiverInput): 필드 별로 useInput 을 통해 분리
